Crate platform

Structs

ApplicationMetadata

AsmDelay

A basic delay implementation.

AsyncFlash

NetSettings

Settings that are used for configuring the network interface to Stabilizer.

SerialSettingsPlatform

SettingsKey

TelemetryClient

The telemetry client for reporting telemetry data over MQTT.

Traits

AppSettings

UnlockFlash

Functions

bootload_dfu

Execute the DFU bootloader stored in system memory.

dfu_flag_is_set

Check if the DFU reboot flag is set, indicating a reboot to DFU is requested.

dfu_reboot

Indicate a reboot to DFU is requested.